ft_printf_12/1_2

+++ ft_printf/ft_printf_utils.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_printf_utils.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jihykim2 <jihykim2@student.42.fr>          +#+  +:+       +#+        */
+/*   By: jihykim2 <jihykim2@student.42seoul.kr>     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/29 14:24:30 by jihykim2          #+#    #+#             */
-/*   Updated: 2022/11/30 22:17:42 by jihykim2         ###   ########.fr       */
+/*   Updated: 2022/12/01 16:46:21 by jihykim2         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -34,7 +34,7 @@ void	print_str(char *str, int *len)
 	}
 }
 
-void	print_nbr(unsigned int num, int *len)
+void	print_nbr(long num, int *len)
 {
 	char	c;
